+++ cmd/adv/main.go
@@ -12,34 +12,22 @@ import (
 	postgres "github.com/thongpham95/adv-tv-backend/internal/adv/postgres"
 	"github.com/thongpham95/adv-tv-backend/internal/adv/repositories"
 	spaces "github.com/thongpham95/adv-tv-backend/internal/adv/spaces"
-	"github.com/thongpham95/adv-tv-backend/internal/adv/utils/errorhandler"
 )
 
 // ClientError is an error whose details to be shared with client.
-
-// Use as a wrapper around the handler functions.
-type rootHandler func(http.ResponseWriter, *http.Request) error
-
-func (fn rootHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	err := fn(w, r) // Call handler function
-	if err == nil {
-		return
-	}
-	// This is where our error handling logic starts.
-	log.Printf("An error occured: %v", err) // Log the error.
-
-	errorhandler.HTTPErrorResponse(w, err)
-}
 
 // Done facilitating error handler
 
 func handleRequests(db *sql.DB, spaceClient *s3.S3) {
 	userRepo := repositories.NewUserRepo(db)
-	videoRepo := repositories.NewVideoRepo(db)
-	controllerHandler := controllers.NewBaseHandler(userRepo, videoRepo, spaceClient)
-	// http.Handle("/login", middlewares.IsAuthorized(rootHandler(controllerHandler.Login)))
-	http.Handle("/login", rootHandler(controllerHandler.Login))
-	http.Handle("/video/upload", middlewares.IsAuthorized(rootHandler(controllerHandler.UploadVideo)))
+	mediaRepo := repositories.NewMediaItem(db)
+	deviceRepo := repositories.NewDevice(db)
+	controllerHandler := controllers.NewBaseHandler(userRepo, mediaRepo, deviceRepo, spaceClient)
+	http.HandleFunc("/login", controllerHandler.Login)                             // POST
+	http.HandleFunc("/signup", controllerHandler.SignUp)                           // POST
+	http.Handle("/device", middlewares.IsAuthorized(controllerHandler.AddDevice))  // POST
+	http.Handle("/media", middlewares.IsAuthorized(controllerHandler.UploadVideo)) // POST
+	http.Handle("/media", middlewares.IsAuthorized(controllerHandler.GetMediaURL)) // GET
 }
 
 func main() {