+++ internal/adv/utils/responsehandler/response.go
@@ -8,6 +8,7 @@ import (
 // HTTPResponse is custom response struct
 type HTTPResponse struct {
 	Success bool        `json:"success"`
+	Message string      `json:"message"`
 	Data    interface{} `json:"data"`
 }
 
@@ -18,10 +19,18 @@ func (r *HTTPResponse) SuccessResponse(w http.ResponseWriter) {
 	json.NewEncoder(w).Encode(r)
 }
 
+// ErrorResponse returns body as JSON
+func (r *HTTPResponse) ErrorResponse(w http.ResponseWriter, status int) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(status)
+	json.NewEncoder(w).Encode(r)
+}
+
 // NewHTTPResponse return a new HTTPResponse
-func NewHTTPResponse(data interface{}) *HTTPResponse {
+func NewHTTPResponse(isSuccess bool, message string, data interface{}) *HTTPResponse {
 	return &HTTPResponse{
-		Success: true,
+		Success: isSuccess,
+		Message: message,
 		Data:    data,
 	}
 }